fix: refactor BackgroundIntent and update widget handling
- Moved BackgroundIntent implementation from BackgroundIntent.swift to TodayTasksWidget.swift for better organization.
- Removed the obsolete BackgroundIntent.swift file.
- Enhanced widget update logic to ensure reliable refresh after task updates.
- Updated project file version to reflect changes.
